IGNITE-21830 * Redundant tests removed
